Finished adding vuex
Finished the general layout of the store and how data is passed around. Also added a few minor updates in some components.
