+++ src/components/Search.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import SearchCheck from '../SearchCheck.js'
+import SearchActions from '../SearchActions.js'
 import Spinner from './Spinner2.vue'
 
 export default {
@@ -35,28 +35,8 @@ export default {
     askWatson () {
       // Start the spinner
       this.loading = true
-      // GET /someUrl
-      this.$http.get('/api', {params: {message: this.query}}).then(response => {
-        console.log(response.body)
-        var comp = response.body.classType
-        // Let router know a search has been performed through this object
-        SearchCheck.searchPerformed(comp)
-        // Create the object that will contain the returned json
-        var whereToGo = {
-          // params: { theResponse: response.body },
-          name: comp
-        }
-        console.log(whereToGo)
-        // Update the store
-        this.$store.commit('newDataRetrieved', {retrievedData: response.body})
-        this.loading = false
-        // Go to this route
-        this.$router.push(whereToGo)
-      }, response => {
-        // error callback, route to error page
-        this.loading = false
-        this.$router.push('/Error')
-      })
+      SearchActions.initialSearch(this, this.query)
+      this.loading = false
     }
   }
 }
